move connectionState to its own file
